# dm-api/src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "本文档说明了每个接口返回、接收的字段信息。并且制定相应的规范作为前后端交流的渠道。",
    "version": "2.0",
    "title": "易授微服务RESTful APIs说明文档",
    "contact": {
      "name": "刘祥德",
      "url": "/",
      "email": "luvletterU@njust.edu.cn"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "分页获取任务列表",
        "operationId": "getListUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "按某属性排序",
            "required": false,
            "type": "string",
            "default": "taskId"
          },
          {
            "name": "direction",
            "in": "query",
            "description": "排序方式",
            "required": false,
            "type": "string",
            "default": "DESC"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«DataMiningTask»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "创建一个数据挖掘任务",
        "operationId": "createUsingPOST_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "miningTaskDto",
            "description": "miningTaskDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MiningTaskDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataMiningTask"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "更新数据挖掘任务的信息",
        "operationId": "updateUsingPUT_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "miningTaskDto",
            "description": "miningTaskDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MiningTaskDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataMiningTask"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}": {
      "get": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "根据ID获取数据挖掘任务的信息",
        "operationId": "getUsingGET_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataMiningTask"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "根据ID删除一个数据挖掘任务",
        "operationId": "deleteByIdUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DataMiningTask"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/algorithms": {
      "get": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "获取该任务配置的算法",
        "operationId": "fetchConfiguredAlgorithmsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Algorithm"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "配置算法",
        "operationId": "configAlgorithmsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "algorithmIds",
            "description": "algorithmIds",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Algorithm"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/collections": {
      "get": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "获取数据集",
        "operationId": "getInvolvedSetsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataSetCollection"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "分配数据集",
        "operationId": "arrangeMiningSetsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "collectionIds",
            "description": "collectionIds",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataSetCollection"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "取消所有数据集与该任务的关联",
        "operationId": "deleteAllRelContainersUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataSetCollection"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/groups": {
      "get": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "获取与此任务相关联的分组",
        "operationId": "groupsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataMiningGroup"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "为多个分组批量分配任务",
        "operationId": "involveGroupsUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "groupIds",
            "description": "groupIds",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataMiningGroup"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "mining-task-controller"
        ],
        "summary": "批量移除任务",
        "operationId": "removeGroupsUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "taskId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "groupIds",
            "description": "groupIds",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DataMiningGroup"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Algorithm": {
      "type": "object",
      "properties": {
        "algorithmId": {
          "type": "integer",
          "format": "int32"
        },
        "algorithmName": {
          "type": "string"
        },
        "interfaceDescription": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/AlgorithmType"
        }
      }
    },
    "AlgorithmType": {
      "type": "object",
      "properties": {
        "chTypeName": {
          "type": "string"
        },
        "enTypeName": {
          "type": "string"
        },
        "typeId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "AreaType": {
      "type": "object",
      "properties": {
        "abbreviation": {
          "type": "string"
        },
        "areaId": {
          "type": "integer",
          "format": "int32"
        },
        "chineseName": {
          "type": "string"
        },
        "englishName": {
          "type": "string"
        }
      }
    },
    "AssociatedTask": {
      "type": "object",
      "properties": {
        "chineseName": {
          "type": "string"
        },
        "englishName": {
          "type": "string"
        },
        "typeId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "AttributeCharacteristic": {
      "type": "object",
      "properties": {
        "abbreviation": {
          "type": "string"
        },
        "charId": {
          "type": "integer",
          "format": "int32"
        },
        "chineseName": {
          "type": "string"
        },
        "englishName": {
          "type": "string"
        }
      }
    },
...
    "DataMiningGroup": {
      "type": "object",
      "properties": {
        "arrangementId": {
          "type": "string"
        },
        "builtTime": {
          "type": "string",
          "format": "date-time"
        },
        "dataMiningTask": {
          "$ref": "#/definitions/DataMiningTask"
        },
        "groupBuilder": {
          "$ref": "#/definitions/Teacher"
        },
        "groupId": {
          "type": "string"
        },
        "groupLeader": {
          "$ref": "#/definitions/Student"
        },
        "groupName": {
          "type": "string"
        }
      }
    },
...
    "DataMiningTask": {
      "type": "object",
      "properties": {
        "actualDuration": {
          "type": "integer",
          "format": "int32"
        },
        "actualFinishTime": {
          "type": "string",
          "format": "date-time"
        },
        "actualStartTime": {
          "type": "string",
          "format": "date-time"
        },
        "algorithms": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Algorithm"
          }
        },
        "arrangementId": {
          "type": "string"
        },
        "builtTime": {
          "type": "string",
          "format": "date-time"
        },
        "plannedDuration": {
          "$ref": "#/definitions/MiningTaskDuration"
        },
        "plannedFinishTime": {
          "type": "string",
          "format": "date-time"
        },
        "plannedStartTime": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "enum": [
            "attentionRequried",
            "completed",
            "inProgress",
            "incomplete",
            "newTask"
          ]
        },
        "taskDescription": {
          "type": "string"
        },
        "taskId": {
          "type": "string"
        },
        "taskName": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "DataSetCharacteristic": {
      "type": "object",
      "properties": {
        "abbreviation": {
          "type": "string"
        },
        "charId": {
          "type": "integer",
          "format": "int32"
        },
        "chineseName": {
          "type": "string"
        },
        "englishName": {
          "type": "string"
        }
      }
    },
    "DataSetCollection": {
      "type": "object",
      "properties": {
        "area": {
          "$ref": "#/definitions/AreaType"
        },
        "associatedTasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssociatedTask"
          }
        },
        "attributeCharacteristics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AttributeCharacteristic"
          }
        },
        "collectionId": {
          "type": "string"
        },
        "collectionName": {
          "type": "string"
        },
        "dataSetCharacteristics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataSetCharacteristic"
          }
        },
        "dataSetFolderPath": {
          "type": "string"
        },
        "dateDonated": {
          "type": "string",
          "format": "date"
        },
        "descriptions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataSetDescription"
          }
        },
        "isMissingValues": {
          "type": "string"
        },
        "numberOfAttributes": {
          "type": "integer",
          "format": "int64"
        },
        "numberOfInstances": {
          "type": "integer",
          "format": "int64"
        },
        "numberOfWebHits": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "DataSetDescription": {
      "type": "object",
      "properties": {
        "descriptionId": {
          "type": "integer",
          "format": "int64"
        },
        "detail": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
    "FavoriteStatus": {
      "type": "object",
      "properties": {
        "chineseValue": {
          "type": "string"
        },
        "englishValue": {
          "type": "string"
        },
        "favoriteId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "MiningTaskDto": {
      "type": "object",
      "properties": {
        "actualFinishTime": {
          "type": "string",
          "format": "date"
        },
        "actualStartTime": {
          "type": "string",
          "format": "date"
        },
        "algorithmIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "arrangementId": {
          "type": "string"
        },
        "builtTime": {
          "type": "string",
          "format": "date"
        },
        "collectionIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "groupIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "plannedFinishTime": {
          "type": "string",
          "format": "date"
        },
        "plannedStartTime": {
          "type": "string",
          "format": "date"
        },
        "statusId": {
          "type": "integer",
          "format": "int32"
        },
        "taskDescription": {
          "type": "string"
        },
        "taskId": {
          "type": "string"
        },
        "taskName": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "MiningTaskDuration": {
      "type": "object",
      "properties": {
        "durationDay": {
          "type": "number",
          "format": "double"
        },
        "durationHour": {
          "type": "number",
          "format": "double"
        },
        "durationMinute": {
          "type": "number",
          "format": "double"
        }
      }
    },
...
    "Page«DataMiningTask»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DataMiningTask"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Sort": {
      "type": "object"
    },
    "Student": {
      "type": "object",
      "properties": {
        "accountStatus": {
          "type": "string",
          "enum": [
            "LOCKED",
            "AVAILABLE"
          ]
        },
        "avatar": {
          "type": "string",
          "format": "byte"
        },
        "birthday": {
          "type": "string",
          "format": "date"
        },
        "className": {
          "type": "string"
        },
        "favorite": {
          "$ref": "#/definitions/FavoriteStatus"
        },
        "gender": {
          "type": "string"
        },
        "grade": {
          "type": "string"
        },
        "introduction": {
          "type": "string"
        },
        "lastLoginTime": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "profession": {
          "type": "string"
        },
        "publicSalt": {
          "type": "string"
        },
        "regionCode": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/StudentStatus"
        },
        "studentId": {
          "type": "string"
        },
        "studentName": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
    "StudentStatus": {
      "type": "object",
      "properties": {
        "chineseValue": {
          "type": "string"
        },
        "englishValue": {
          "type": "string"
        },
        "statusId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Teacher": {
      "type": "object",
      "properties": {
        "accountStatus": {
          "type": "string",
          "enum": [
            "LOCKED",
            "AVAILABLE"
          ]
        },
        "avatar": {
          "type": "string",
          "format": "byte"
        },
        "birthday": {
          "type": "string",
          "format": "date"
        },
        "gender": {
          "type": "string"
        },
        "introduction": {
          "type": "string"
        },
        "lastLoginTime": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "publicSalt": {
          "type": "string"
        },
        "regionCode": {
          "type": "string"
        },
        "teacherId": {
          "type": "string"
        },
        "teacherName": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
  }
}